ai-service/requirements.txt
apscheduler==3.11.1
common==0.1.2
fastapi>=0.115.0
langchain_openai==1.1.7
litellm==1.81.6
loguru==0.7.2
openai==2.16.0
pydantic==2.12.5
pydantic_settings==2.12.0
python-dotenv==1.2.1
uvicorn>=0.30.0
opik>=0.2.0
pgvector
sentence-transformers
tiktoken
pypdf
python-docx
psycopg[binary]
